# frontend/packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ArchivedFilter": {
        "description": "Whether archived clients appear — FR-M1-014.\n\n🔒 DB §22.2: \"Archived clients — excluded from lists, search, entitlement\ncount, all messaging.\" So ``EXCLUDE`` is the default and the partial indexes\ncarry the predicate.\n\n⚠️ ``ONLY`` exists because a practitioner who archived somebody by mistake\nneeds to find them to restore them (EC-M1-02), and a client who cannot be\nfound cannot be restored. Without it the archive is a one-way door.",
        "enum": [
          "exclude",
          "only",
          "include"
        ],
        "title": "ArchivedFilter",
        "type": "string"
      },
...
      "ClientListResponse": {
        "description": "API §5.1's collection envelope.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ClientListItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/ClientPageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "ClientListResponse",
        "type": "object"
      },
...
      "ClientStage": {
        "description": "The lifecycle — DB §5.2, M1.4. 🟡 Values PROPOSED pending OD-01.\n\n🔒 One entity, seven stages. A ``lead`` and an ``active`` client are the same\nrow; the stage is what differs. That is what makes AC-M1-003 (\"converting a\nlead retains the original record, its identifier, and all prior history\")\ntrue by construction rather than by careful copying.",
        "enum": [
          "lead",
          "contacted",
          "consultation_scheduled",
          "active",
          "paused",
          "churned",
          "archived"
        ],
        "title": "ClientStage",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Practice management for Indian nutrition practitioners.\n\n🔒 Realm-segmented paths (ADR-A01): `/app` practitioner, `/portal` client, `/admin` operator, `/public` unauthenticated.",
    "title": "WellnessCRM V2 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/app/clients": {
      "get": {
        "description": "The client list — FR-M1-021/022, NFR-005 (≤300 ms).\n\n🔒 **A practitioner sees only their own and their assigned clients**\n(AC-M1-006); an owner sees the tenant (FR-M0-017). Applied as a predicate\ninside the query, because a list filtered after the fact would page short and\nleak totals.\n\n⚠️ **Stages are OR, tags are AND.** Ticking two stages means \"either\" — the\nnatural reading of a status filter. Ticking two tags means \"both\", because\ntags are how a caseload is *narrowed*: \"PCOS or post-natal\" would return a\nlonger list than the practitioner started with, which is the opposite of what\nchoosing a second filter is for.\n\n⚠️ Archived clients are excluded by default (FR-M1-014, DB §22.2).\n``archived=only`` exists so a client archived by mistake can be found and\nrestored — without it the archive is a one-way door.",
        "operationId": "clientsList",
        "parameters": [
          {
            "description": "Search name, email or the last digits of a mobile (FR-M1-021).",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search name, email or the last digits of a mobile (FR-M1-021).",
              "title": "Q"
            }
          },
          {
            "description": "Repeat for OR within the field — API §6.2.",
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ClientStage"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeat for OR within the field — API §6.2.",
              "title": "Stage"
            }
          },
          {
            "description": "Repeat to require ALL named tags — see below.",
            "in": "query",
            "name": "tag_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeat to require ALL named tags — see below.",
              "title": "Tag Id"
            }
          },
          {
            "in": "query",
            "name": "owner_user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner User Id"
            }
          },
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ArchivedFilter",
              "default": "exclude"
            }
          },
          {
            "description": "`name`, `recent_activity` or `created`; `-` prefix for descending.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`name`, `recent_activity` or `created`; `-` prefix for descending.",
              "title": "Sort"
            }
          },
          {
            "description": "Opaque, from a previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque, from a previous page.",
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Adds a COUNT — API §6.1.",
            "in": "query",
            "name": "include_total",
            "required": false,
            "schema": {
              "default": false,
              "description": "Adds a COUNT — API §6.1.",
              "title": "Include Total",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List, search and filter clients",
        "tags": [
          "clients"
        ]
      },
      "post": {
        "description": "Create a client or lead — FR-M1-004, AC-M1-001.\n\n🔒 Not metered. FR-M1-003 keeps leads unmetered at every stage before\n``active``, and EC-M2-06 requires a tenant at their limit to keep accepting\nthem. The entitlement binds on the transition to ``active`` (Slice B).",
        "operationId": "clientsCreate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a client",
        "tags": [
          "clients"
        ]
      }
    },
...
    "/api/v1/app/clients/{client_id}": {
      "get": {
        "description": "One client by id — API §7.1.\n\n⚠️ Archived clients are returned. AC-M1-007 requires archiving to remove them\nfrom default *views* without deleting anything, and restoring one requires\nbeing able to read it first.",
        "operationId": "clientsRead",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a client",
        "tags": [
          "clients"
        ]
      },
      "patch": {
        "description": "Apply a partial edit — API §4.4.\n\n🔒 ``If-Match`` is **required**. Two practitioners editing one client is\nroutine in a clinic, and without a precondition the second save silently\ndiscards the first one's work with no error and no trace.",
        "operationId": "clientsUpdate",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Client Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update a client",
        "tags": [
          "clients"
        ]
      }
    },
...
  }
}